chore: `metMinutes` are now returned as integers (#142)

+++ ios/Analytics/Sources/ActivityMeasure.swift
@@ -3,6 +3,6 @@ import Foundation
 public struct ActivityMeasure: Codable {
 
     public let seconds: TimeInterval
-    public let metMinutes: Float
+    public let metMinutes: Int
 
 }
